kobopatch-livepatch: Remove kobopatch-livepatch
Although it worked most of the time (it was kind of buggy), it's not really
practical to use on a regular basis due to its performance and its limitations
with regards to when things are initialized.

If I choose to work on this again, it would be from scratch with a different
approach.