Fix parsing of epub3 metadata
Use a QString for 'name' instead of a QStringRef to ensure its value
doesn't change even after calling QXmlStreamReader::readElementText(),
which may update the content of the internal buffer that 'name' was
referring to.  Without this, later comparisons against 'name' may fail
as it no longer points to the value of the property attribute, but
rather a substring of the <meta> element's text.